requirements.txt
Babel==2.12.1
Pillow==9.5.0
